+++ CHANGELOG.md
@@ -14,6 +14,10 @@ If you like this project and find it useful, please consider giving it a star on
 
 ### Added
 
+- [frontend]: Bumped `frontend` version to 3.1.0. Fully typed with the backend.
+- [frontend]: Created an autonomous component `MbfTable` to remove `react-table`. Features: unique UI for all tables with integrated column sorting and column selection.
+- [frontend]: Use MbfTable for Plugins, Devices tables.
+
 ### Changed
 
 - [matter]: Bumped `matter.js` version to 0.15.4.